+++ Modules/Intent.Modules.Application.MediatR.CRUD.Tests/Templates/Owner/GetAllPaginationQueryHandlerTests/GetAllPaginationQueryHandlerTestsTemplatePartial.cs
@@ -5,31 +5,27 @@ using Intent.Engine;
 using Intent.Modelers.Domain.Api;
 using Intent.Modelers.Services.Api;
 using Intent.Modelers.Services.CQRS.Api;
-using Intent.Modules.Application.MediatR.CRUD.CrudStrategies;
 using Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Assertions.AssertionClass;
-using Intent.Modules.Application.MediatR.Templates;
-using Intent.Modules.Application.MediatR.Templates.QueryModels;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Constants;
-using Intent.Modules.Entities.Repositories.Api.Templates;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
 
-namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Nested.NestedGetAllQueryHandlerTests
+namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Owner.GetAllPaginationQueryHandlerTests
 {
     [IntentManaged(Mode.Fully, Body = Mode.Merge)]
-    public partial class NestedGetAllQueryHandlerTestsTemplate : CSharpTemplateBase<QueryModel>, ICSharpFileBuilderTemplate
+    public partial class GetAllPaginationQueryHandlerTestsTemplate : CSharpTemplateBase<QueryModel>, ICSharpFileBuilderTemplate
     {
-        public const string TemplateId = "Intent.Application.MediatR.CRUD.Tests.Nested.NestedGetAllQueryHandlerTests";
+        public const string TemplateId = "Intent.Application.MediatR.CRUD.Tests.Owner.GetAllPaginationQueryHandlerTests";
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
-        public NestedGetAllQueryHandlerTestsTemplate(IOutputTarget outputTarget, QueryModel model) : base(TemplateId, outputTarget, model)
+        public GetAllPaginationQueryHandlerTestsTemplate(IOutputTarget outputTarget, QueryModel model) : base(TemplateId, outputTarget, model)
         {
             AddNugetDependency(NugetPackages.AutoFixture);
             AddNugetDependency(NugetPackages.FluentAssertions);
@@ -48,9 +44,8 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Nested.NestedG
                 {
                     AddUsingDirectives(file);
                     Facade.AddHandlerConstructorMockUsings();
-
+                    
                     var priClass = file.Classes.First();
-
                     priClass.AddConstructor(ctor =>
                     {
                         ctor.AddStatements(Facade.GetAutoMapperProfilesAndAddBackendField(priClass));
@@ -61,31 +56,27 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Nested.NestedG
                         method.Static();
                         method.AddStatements(Facade.Get_InitialAutoFixture_TestDataStatements(
                             includeVarKeyword: true));
-                        method.AddStatements(Facade.Get_SingleDomainEntity_TestDataStatements(
-                            queryTargetDomain: QueryTargetDomain.NestedEntity,
-                            dataReturn: QueryTestDataReturn.AggregateDomain,
-                            includeVarKeyword: true,
-                            nestedEntitiesAreEmpty: false));
-                        method.AddStatements(Facade.Get_InitialAutoFixture_TestDataStatements(
-                            includeVarKeyword: false));
-                        method.AddStatements(Facade.Get_SingleDomainEntity_TestDataStatements(
-                            queryTargetDomain: QueryTargetDomain.NestedEntity,
-                            dataReturn: QueryTestDataReturn.AggregateDomain,
-                            includeVarKeyword: false,
-                            nestedEntitiesAreEmpty: true));
+                        method.AddStatements(Facade.Get_ManyAggregateDomainEntities_TestDataStatements(5));
+                        method.AddStatements(Facade.Get_ManyAggregateDomainEntities_TestDataStatements(0));
                     });
-
+                    
                     priClass.AddMethod("Task", $"Handle_WithValidQuery_Retrieves{Facade.PluralTargetDomainName}", method =>
                     {
                         method.Async();
                         method.AddAttribute("Theory");
                         method.AddAttribute("MemberData(nameof(GetSuccessfulResultTestData))");
-                        method.AddParameter(Facade.AggregateOwnerDomainTypeName, "existingOwnerEntity");
+                        method.AddParameter($"List<{Facade.TargetDomainTypeName}>", "testEntities");
 
                         method.AddStatement("// Arrange");
                         method.AddStatements(Facade.GetNewQueryAutoFixtureInlineStatements("testQuery"));
+                        method.AddStatements(Facade.GetPaginationQuerySetup("testQuery", pageNo: 1, pageSize: 5));
                         method.AddStatements(Facade.GetQueryHandlerConstructorParameterMockStatements());
-                        method.AddStatements(Facade.GetDomainAggregateOwnerRepositoryFindByIdMockingStatements("testQuery", "existingOwnerEntity", QueryHandlerFacade.MockRepositoryResponse.ReturnDomainVariable));
+                        method.AddStatements(Facade.GetMockedPaginatedResults("fetchedResults", "testEntities"));
+                        method.AddStatements(Facade.GetDomainRepositoryFindAllMockingStatements(
+                            entitiesVarName: "fetchedResults", 
+                            repositoryVarName: Facade.DomainAggregateRepositoryVarName, 
+                            pageNo: 1, 
+                            pageSize: 5));
                         method.AddStatements(Facade.GetQueryHandlerConstructorSutStatement());
 
                         method.AddStatement(string.Empty);
@@ -94,9 +85,9 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Nested.NestedG
 
                         method.AddStatement(string.Empty);
                         method.AddStatement("// Assert");
-                        method.AddStatements(Facade.Get_AggregateOwner_AssertionComparingHandlerResultsWithExpectedResults($"existingOwnerEntity.{Facade.OwnerToCompositeNavigationPropertyName}"));
+                        method.AddStatements(Facade.Get_Aggregate_AssertionComparingHandlerResultsWithExpectedResults("fetchedResults"));
                     });
-
+                    
                     AddAssertionMethods();
                 });
         }
@@ -119,27 +110,19 @@ namespace Intent.Modules.Application.MediatR.CRUD.Tests.Templates.Nested.NestedG
 
         private void AddAssertionMethods()
         {
-            var dtoModel = Model.TypeReference.Element.AsDTOModel();
-
-            if (dtoModel?.Mapping?.Element?.IsClassModel() != true)
+            if (!Model.IsPaginationQueryMapping())
             {
                 return;
             }
+            
+            var domainModel = Model.GetPaginatedClassModel();
+            var templateInstance = ExecutionContext.FindTemplateInstance<ICSharpFileBuilderTemplate>(
+                TemplateDependency.OnModel(AssertionClassTemplate.TemplateId, domainModel));
 
-            var nestedDomainElement = dtoModel.Mapping.Element.AsClassModel();
-            var ownerDomainElement = nestedDomainElement.GetNestedCompositionalOwner();
-            var template = ExecutionContext.FindTemplateInstance<ICSharpFileBuilderTemplate>(
-                TemplateDependency.OnModel(AssertionClassTemplate.TemplateId, ownerDomainElement));
-            if (template == null)
-            {
-                return;
-            }
-
-            template.AddAssertionMethods(template.CSharpFile.Classes.First(), nestedDomainElement, dtoModel, true);
+            templateInstance?.AddPaginationAssertionMethod(templateInstance.CSharpFile.Classes.First(), Model, domainModel);
         }
 
-        [IntentManaged(Mode.Fully)]
-        public CSharpFile CSharpFile { get; }
+        [IntentManaged(Mode.Fully)] public CSharpFile CSharpFile { get; }
 
         [IntentManaged(Mode.Fully)]
         protected override CSharpFileConfig DefineFileConfig()