+++ Modules/Intent.Modules.Application.MediatR.CRUD.Tests/release-notes.md
@@ -5,6 +5,7 @@
 - Improvement: Now catering for composite keys in Command / Query handlers.
 - Improvement: Now generating tests for Commands projecting to Constructors and Operations.
 - Improvement: Assertion classes will now only generate if being used by Unit Tests.
+- Improvement: Now generating tests for Queries that are setup for pagination.
 - Fixed: Get query handler tests would cause a crash in some cases when trying to generate.
 - Fixed: Inaccurately named unit tests.
 - Fixed: Assertions classes had a wrong comparison input on `HaveSameCount()` calls.